# services/orion-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ORION API",
    "version": "0.2"
  },
  "paths": {
...
    "/ask": {
      "post": {
        "summary": "Ask",
        "operationId": "ask_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/chat": {
      "post": {
        "summary": "Chat",
        "operationId": "chat_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/workflows": {
      "get": {
        "summary": "List Workflows",
        "description": "List all registered workflows with current execution status.",
        "operationId": "list_workflows_workflows_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkflowResponse"
                  },
                  "type": "array",
                  "title": "Response List Workflows Workflows Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Workflow",
        "description": "Create a new workflow from a node definition or from a built-in template.\nIf template_id is provided, nodes are generated from the template.",
        "operationId": "create_workflow_workflows_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkflowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AskRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt",
            "description": "The query prompt sent to ORION",
            "examples": [
              "Hello"
            ]
          },
          "confirmed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed",
            "description": "Whether a potentially destructive action is confirmed",
            "default": false
          },
          "confirmation_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmation Token",
            "description": "Correlation token confirming a blocked action"
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "AskRequest"
      },
      "AskResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the request processing succeeded"
          },
          "intent": {
            "type": "string",
            "title": "Intent",
            "description": "The classified intent of the user prompt"
          },
          "response": {
            "type": "string",
            "title": "Response",
            "description": "The response generated by the system"
          },
          "tool_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Used",
            "description": "The tool executed, if any"
          },
          "tool_output": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Output",
            "description": "The tool execution output, if any"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "description": "The session identifier"
          },
          "execution_time_ms": {
            "type": "number",
            "title": "Execution Time Ms",
            "description": "The execution duration in milliseconds"
          },
          "telemetry": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TelemetryDetail"
              },
              {
                "type": "null"
              }
            ],
            "description": "Generative telemetry statistics"
          },
          "confirmation_required": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmation Required",
            "description": "Indicates if user confirmation is required to proceed",
            "default": false
          },
          "confirmation_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmation Token",
            "description": "Unique token for confirming this action"
          }
        },
        "type": "object",
        "required": [
          "success",
          "intent",
          "response",
          "session_id",
          "execution_time_ms"
        ],
        "title": "AskResponse"
      },
      "ChatRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt",
            "description": "The query prompt sent to ORION"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "The session ID to persist history"
          },
          "stream": {
            "type": "boolean",
            "title": "Stream",
            "description": "Set to true to stream tokens dynamically",
            "default": false
          },
          "confirmed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed",
            "description": "Whether a potentially destructive action is confirmed",
            "default": false
          },
          "confirmation_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmation Token",
            "description": "Correlation token confirming a blocked action"
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "ChatRequest"
      },
      "ConditionConfigRequest": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "boolean"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params",
            "default": {}
          },
          "on_true": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "On True"
          },
          "on_false": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "On False"
          }
        },
        "type": "object",
        "title": "ConditionConfigRequest"
      },
      "CreateWorkflowRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "flow_type": {
            "type": "string",
            "title": "Flow Type",
            "default": "SEQUENTIAL"
          },
          "template_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "default": []
          },
          "timeout_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Seconds"
          },
          "max_retries": {
            "type": "integer",
            "title": "Max Retries",
            "default": 0
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/WorkflowNodeRequest"
            },
            "type": "array",
            "title": "Nodes",
            "default": []
          },
          "variables": {
            "additionalProperties": true,
            "type": "object",
            "title": "Variables",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateWorkflowRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TelemetryDetail": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model",
            "description": "Model provider key"
          },
          "prompt_tokens": {
            "type": "integer",
            "title": "Prompt Tokens",
            "description": "Tokens used for prompt",
            "default": 0
          },
          "completion_tokens": {
            "type": "integer",
            "title": "Completion Tokens",
            "description": "Tokens used for completion",
            "default": 0
          },
          "total_tokens": {
            "type": "integer",
            "title": "Total Tokens",
            "description": "Total tokens used",
            "default": 0
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Error details, if any"
          }
        },
        "type": "object",
        "required": [
          "model"
        ],
        "title": "TelemetryDetail"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkflowNodeRequest": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "flow_type": {
            "type": "string",
            "title": "Flow Type",
            "default": "SEQUENTIAL"
          },
          "depends_on": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Depends On",
            "default": []
          },
          "on_success": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "On Success"
          },
          "on_failure": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "On Failure"
          },
          "inputs": {
            "additionalProperties": true,
            "type": "object",
            "title": "Inputs",
            "default": {}
          },
          "max_retries": {
            "type": "integer",
            "title": "Max Retries",
            "default": 0
          },
          "delay_seconds": {
            "type": "number",
            "title": "Delay Seconds",
            "default": 0.0
          },
          "loop_max": {
            "type": "integer",
            "title": "Loop Max",
            "default": 1
          },
          "condition": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConditionConfigRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "type"
        ],
        "title": "WorkflowNodeRequest"
      },
      "WorkflowNodeResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "flow_type": {
            "type": "string",
            "title": "Flow Type"
          },
          "depends_on": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Depends On"
          },
          "on_success": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "On Success"
          },
          "on_failure": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "On Failure"
          },
          "inputs": {
            "additionalProperties": true,
            "type": "object",
            "title": "Inputs"
          },
          "outputs": {
            "additionalProperties": true,
            "type": "object",
            "title": "Outputs"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "retry_count": {
            "type": "integer",
            "title": "Retry Count"
          },
          "loop_count": {
            "type": "integer",
            "title": "Loop Count"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "type",
          "status",
          "flow_type",
          "depends_on",
          "on_success",
          "on_failure",
          "inputs",
          "outputs",
          "error",
          "started_at",
          "finished_at",
          "retry_count",
          "loop_count",
          "metadata"
        ],
        "title": "WorkflowNodeResponse"
      },
      "WorkflowResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "flow_type": {
            "type": "string",
            "title": "Flow Type"
          },
          "template_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "nodes": {
            "additionalProperties": {
              "$ref": "#/components/schemas/WorkflowNodeResponse"
            },
            "type": "object",
            "title": "Nodes"
          },
          "variables": {
            "additionalProperties": true,
            "type": "object",
            "title": "Variables"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "status",
          "flow_type",
          "template_id",
          "tags",
          "nodes",
          "variables",
          "created_at",
          "updated_at",
          "error",
          "metadata"
        ],
        "title": "WorkflowResponse"
      },
...
    }
  }
}